+++ R/corpus.R
@@ -155,37 +155,52 @@ corpus_install <- function(pkg = NULL, repo = "https://PolMine.github.io/drat/",
 
       if (verbose) cli_rule("Resolve DOI")
       if (verbose) cli_process_start("get Zenodo record referred to by DOI")
-      tryCatch(
-        zenodo_record <- ZenodoManager$new()$getRecordByDOI(doi = doi),
-        error = function(e){
-          if (verbose){
-            cli_process_failed()
-          } else {
-            cli_alert_danger(sprintf("'no Zenodo record found for DOI '%s'", doi))
-          }
-        }
-      )
-      if (!exists("zenodo_record")){
-        return(invisible(FALSE)) # unlikely scenario that can only result from error detected by tryCatch()
-      } else if (is.null(zenodo_record)){
-        if (verbose) cli_alert_danger(sprintf("no Zenodo record found for DOI '%s'", doi))
-        return(invisible(FALSE))
-      } else {
-        if (verbose) cli_process_done()
-      }
-
-      zenodo_files <- sapply(zenodo_record[["files"]], function(x) x[["links"]][["download"]])
-      tarball <- grep("^.*?_(v|)\\d+\\.\\d+\\.\\d+\\.tar\\.gz$", zenodo_files, value = TRUE)
-
-      if (length(tarball) > 1L && isTRUE(ask)){
-        userchoice <- utils::menu(
-          choices = basename(tarball),
-          title = "Several tarballs assumed to contain an indexed corpus are available. Which tarball shall be downloaded?"
+      # tryCatch(
+      #   zenodo_record <- ZenodoManager$new()$getRecordByDOI(doi = doi),
+      #   error = function(e){
+      #     if (verbose){
+      #       cli_process_failed()
+      #     } else {
+      #       cli_alert_danger(sprintf("'no Zenodo record found for DOI '%s'", doi))
+      #     }
+      #   }
+      # )
+      # 
+      # if (!exists("zenodo_record")){
+      #   return(invisible(FALSE)) # unlikely scenario that can only result from error detected by tryCatch()
+      # } else if (is.null(zenodo_record)){
+      #   if (verbose) cli_alert_danger(sprintf("no Zenodo record found for DOI '%s'", doi))
+      #   return(invisible(FALSE))
+      # } else {
+      #   if (verbose) cli_process_done()
+      # }
+      # 
+      # zenodo_files <- sapply(zenodo_record[["files"]], function(x) x[["links"]][["download"]])
+      # tarball <- grep("^.*?_(v|)\\d+\\.\\d+\\.\\d+\\.tar\\.gz$", zenodo_files, value = TRUE)
+      # 
+      # if (length(tarball) > 1L && isTRUE(ask)){
+      #   userchoice <- utils::menu(
+      #     choices = basename(tarball),
+      #     title = "Several tarballs assumed to contain an indexed corpus are available. Which tarball shall be downloaded?"
+      #   )
+      #   tarball <- tarball[userchoice]
+      # }
+      
+      tarball <- zenodo_get_tarballurl(url = doi)
+      if (verbose) cli_process_done()
+      
+      if (verbose) cli_alert_info(
+        sprintf(
+          "tarball to be downloaded: %s",
+          col_blue(basename(unname(tarball)))
         )
-        tarball <- tarball[userchoice]
-      }
-      if (verbose) cli_alert_info(sprintf("tarball to be downloaded: %s", col_blue(basename(tarball))))
-      zenodo_file_record <- zenodo_record[["files"]][[which(zenodo_files == tarball)]]
+      )
+      
+      # zenodo_file_record <- zenodo_record[["files"]][[which(zenodo_files == tarball)]]
+      
+      if (!is.null(checksum)) checksum <- names(tarball)[1]
+      tarball <- unname(tarball)
+      
     }
   }
 
@@ -289,12 +304,12 @@ corpus_install <- function(pkg = NULL, repo = "https://PolMine.github.io/drat/",
           "download corpus tarball {col_cyan('%s')} ... done", basename(tarball)
         )
       )
-      if (exists("zenodo_file_record")){
-        if (!is.null(checksum)){
-          if (verbose) cli_alert_warning("argument checksum is not NULL but md5 checksum can be derived from Zenodo record - using checksum issued by Zenodo")
-        }
-        checksum <- zenodo_file_record[["checksum"]]
-      }
+      # if (exists("zenodo_file_record")){
+      #   if (!is.null(checksum)){
+      #     if (verbose) cli_alert_warning("argument checksum is not NULL but md5 checksum can be derived from Zenodo record - using checksum issued by Zenodo")
+      #   }
+      #   checksum <- zenodo_file_record[["checksum"]]
+      # }
       if (isFALSE(is.null(checksum))){
         if (verbose){
           msg <- sprintf("check md5 checksum for tarball %s (expected %s)", basename(tarball), checksum)
@@ -391,9 +406,9 @@ corpus_install <- function(pkg = NULL, repo = "https://PolMine.github.io/drat/",
     
     # Ask user before overwriting existing corpus ----------------
     for (corpus in corpora){
-      if (exists("zenodo_record")){
-        version <- zenodo_record[["metadata"]][["version"]]
-      } else {
+      # if (exists("zenodo_record")){
+      #   version <- zenodo_record[["metadata"]][["version"]]
+      # } else {
         rf <- registry_file_parse(
           corpus = corpus,
           registry_dir = tmp_registry_dir
@@ -405,7 +420,7 @@ corpus_install <- function(pkg = NULL, repo = "https://PolMine.github.io/drat/",
         } else {
           version <- "unknown"
         }
-      }
+      # }
       
       if (tolower(corpus) %in% list.files(cwb_dirs[["registry_dir"]])){
         regdata <- registry_file_parse(